requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.23.3
pydantic==2.8.2
pytest==8.3.2
